Selection for instance that's not required nor set
An instance and alias are very similar, both using a select where None
is allowed. But for an instance the value for None is the empty string
whereas the value for None for an alias is the class being created by
the form. Added more checks in the code for selection validation to make
sure that the required field must be set when the value is the empty
string will trigger an error, and a None selection for an alias is the
only thing that displays a Warning.