ompound field was popping up dialog
Not sure if this was for debugging or if trying to inform the user that
they are missing some fields, but a dialog was popping up as long as
the compound form was still not valid. Made it difficult to use the
form. Removed the alert from the form-common.js script so that the
dialog wouldn't pop up any longer.